Impliment post's details
