+++ app/graphql/types/comment_type.rb
@@ -8,5 +8,6 @@ module Types
     field :post_id, Integer, null: false
     field :created_at, GraphQL::Types::ISO8601DateTime, null: false
     field :updated_at, GraphQL::Types::ISO8601DateTime, null: false
+    field :user, Types::UserType, null: false
   end
 end
